+++ server/src/game_server/gameObjects/obstacle.ts
@@ -7,7 +7,7 @@ import { type Player } from "./player.ts";
 import { ServerGameObject } from "../others/gameObject.ts";
 import { GameItem } from "common/scripts/definitions/utils.ts";
 import { LootTables } from "common/scripts/definitions/maps/base.ts";
-import { Explosions } from "common/scripts/definitions/explosions.ts";
+import { Explosions } from "../../../../common/scripts/definitions/objects/explosions.ts";
 
 export class Obstacle extends ServerGameObject{
     stringType:string="obstacle"