+++ server/src/game_server/others/game.ts
@@ -11,7 +11,7 @@ import { GameMap, generation } from "./map.ts"
 import { Explosion } from "../gameObjects/explosion.ts";
 import { DefaultGamemode, Gamemode } from "./gamemode.ts";
 import { BulletDef, GameItem } from "common/scripts/definitions/utils.ts";
-import { ExplosionDef } from "common/scripts/definitions/explosions.ts";
+import { ExplosionDef } from "../../../../common/scripts/definitions/objects/explosions.ts";
 import { ProjectileDef } from "common/scripts/definitions/projectiles.ts";
 import { Projectile } from "../gameObjects/projectile.ts";
 import { ServerGameObject } from "./gameObject.ts";